Extract privileged access into contect manager.
