+++ src/plone/restapi/services/workflow/transition.py
@@ -1,19 +1,15 @@
 # -*- coding: utf-8 -*-
 
-from AccessControl import getSecurityManager
-from AccessControl.SecurityManagement import newSecurityManager
-from AccessControl.SecurityManagement import setSecurityManager
-from AccessControl.User import UnrestrictedUser as BaseUnrestrictedUser
 from DateTime import DateTime
 from plone.restapi.deserializer import json_body
 from plone.restapi.interfaces import IDeserializeFromJson
 from plone.restapi.serializer.converters import json_compatible
 from plone.restapi.services import Service
+from plone.restapi.services.workflow.utils import elevated_privileges
 from Products.CMFCore.interfaces import IFolderish
 from Products.CMFCore.utils import getToolByName
 from Products.CMFCore.WorkflowCore import WorkflowException
 from zExceptions import BadRequest
-from zope.component import getMultiAdapter
 from zope.component import queryMultiAdapter
 from zope.i18n import translate
 from zope.interface import alsoProvides
@@ -23,14 +19,6 @@ from zope.publisher.interfaces import NotFound
 
 import plone.protect.interfaces
 import six
-
-
-class UnrestrictedUser(BaseUnrestrictedUser):
-    """Unrestricted user that still has an id."""
-
-    def getId(self):
-        """Return the ID of the user."""
-        return self.getUserName()
 
 
 @implementer(IPublishTraverse)
@@ -90,40 +78,30 @@ class WorkflowTransition(Service):
             self.request.response.setStatus(400)
             return dict(error=dict(type="Bad Request", message=str(e)))
 
-        sm = getSecurityManager()
-        portal = getMultiAdapter(
-            (self.context, self.request), name="plone_portal_state"
-        ).portal()
-        try:
-            tmp_user = UnrestrictedUser(
-                sm.getUser().getId(), "", ("manage", "Manager"), ""
-            )
-            tmp_user = tmp_user.__of__(portal.acl_users)
-            newSecurityManager(None, tmp_user)
-            history = self.wftool.getInfoFor(self.context, "review_history")
-            action = history[-1]
-            if six.PY2:
-                action["title"] = self.context.translate(
-                    self.wftool.getTitleForStateOnType(
-                        action["review_state"], self.context.portal_type
-                    ).decode("utf8")
-                )
-            else:
-                action["title"] = self.context.translate(
-                    self.wftool.getTitleForStateOnType(
-                        action["review_state"], self.context.portal_type
+        with elevated_privileges(self.context):
+            try:
+                history = self.wftool.getInfoFor(self.context, "review_history")
+                action = history[-1]
+                if six.PY2:
+                    action["title"] = self.context.translate(
+                        self.wftool.getTitleForStateOnType(
+                            action["review_state"], self.context.portal_type
+                        ).decode("utf8")
+                    )
+                else:
+                    action["title"] = self.context.translate(
+                        self.wftool.getTitleForStateOnType(
+                            action["review_state"], self.context.portal_type
+                        )
+                    )
+            except WorkflowException as e:
+                self.request.response.setStatus(400)
+                action = dict(
+                    error=dict(
+                        type="WorkflowException",
+                        message=translate(str(e), context=self.request),
                     )
                 )
-        except WorkflowException as e:
-            self.request.response.setStatus(400)
-            action = dict(
-                error=dict(
-                    type="WorkflowException",
-                    message=translate(str(e), context=self.request),
-                )
-            )
-        finally:
-            setSecurityManager(sm)
 
         return json_compatible(action)
